Reactive code should *never* be blocked because it defeats the purpose of reactive: the idea of reactive is exactly not blocking your thread, so it can handle other requests while yours is waiting for blocking resources (in this case, the database).  And this is what we were doing when we were waiting (specially `indefinitely`).  Another sign that something was not 100% was that we were receiving an `Uni` from the repository, working with it and, at the return statement, creating a new uni.  The solution for that is chaining everything in one "big" chain. This way, while the thread is waiting for the DB to answer, it can work on other things.
 Further reading:
https://vertx.io/docs/vertx-core/java/#golden_rule
and
https://en.wikipedia.org/wiki/C10k_problem

+++ src/main/java/com/magalhaes/andre/resource/SalesmanResource.java
@@ -37,7 +37,7 @@ public class SalesmanResource {
     @Path("/quantity")
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public Uni<Map<String, Double>> listSalesmanBySaleQuantity(){ return service.listSalesmanBySaleQuantity(); }
+    public Uni<Map<String, Long>> listSalesmanBySaleQuantity(){ return service.listSalesmanBySaleQuantity(); }
 
     @Path("/total")
     @GET
